MAINT prepare refactoring
